# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hariri Foundation API",
        "description": "API documentation for Users Management",
        "contact": {
            "email": "support@haririfoundation.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/project-employees": {
            "get": {
                "tags": [
                    "Project Employees"
                ],
                "summary": "Get all project employees or a specific one by ID",
                "operationId": "c20792834fafb1b5ca989c649b2d0fbe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Optional UUID to fetch a specific project employee",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project Employees retrieved successfully"
                    },
                    "404": {
                        "description": "Project Employee not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Project Employees"
                ],
                "summary": "Create a new project employee",
                "operationId": "6e7925fac973413a6b4a203a924d6acb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "program_id",
                                    "employee_id"
                                ],
                                "properties": {
                                    "program_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "123e4567-e89b-12d3-a456-426614174000"
                                    },
                                    "employee_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "321e6547-e89b-12d3-a456-426614174000"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Project coordination duties"
                                    },
                                    "external_id": {
                                        "type": "string",
                                        "example": "EXT-10023"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project Employee created successfully"
                    },
                    "400": {
                        "description": "Invalid input data"
                    }
                }
            }
        },
        "/api/project-employees/{id}": {
            "put": {
                "tags": [
                    "Project Employees"
                ],
                "summary": "Update a project employee",
                "operationId": "1d76f3222108ca8e25113a81e4e6739b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project Employee UUID to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "program_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "employee_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "external_id": {
                                        "type": "string",
                                        "example": "EXT-9999"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project Employee updated successfully"
                    },
                    "404": {
                        "description": "Project Employee not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Project Employees"
                ],
                "summary": "Delete a project employee",
                "operationId": "538a820321af365d7c6ebe8eea16405e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of project employee to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project Employee deleted successfully"
                    },
                    "404": {
                        "description": "Project Employee not found"
                    }
                }
            }
        },
...
    },
...
}